fix: import PracticeStats in Mappers, ScenePicker setValue import, dashboard fraction type
Three CI-discovered compile errors:
  - Mappers.kt referenced PracticeStats without importing it (Book/DailyScore/DialogueScene got imported in a previous fix but PracticeStats was missed)
  - FreeDialogueScreen.ScenePicker used fully-qualified androidx.compose.runtime.remember + mutableStateOf inline, which broke the  delegate (needs setValue in scope)
  - DashboardScreen.ScoreProgress passed (Double/100.0).coerceIn(0.0, 1.0) to fillMaxWidth(fraction = ...) which expects Float; added .toFloat() coercion

+++ android/app/src/main/java/com/app/english/ui/dashboard/DashboardScreen.kt
@@ -384,7 +384,7 @@ private fun ScoreProgress(value: Double) {
                 shape = RoundedCornerShape(6.dp),
                 color = color,
                 modifier = Modifier
-                    .fillMaxWidth(fraction = (value / 100.0).coerceIn(0.0, 1.0))
+                    .fillMaxWidth(fraction = (value / 100.0).toFloat().coerceIn(0f, 1f))
                     .fillMaxSize()
             ) {}
         }
